+++ app/controllers/AnalyzerCtrl.scala
@@ -2,15 +2,13 @@ package controllers
 
 import javax.inject.Inject
 
-import scala.annotation.implicitNotFound
-import scala.concurrent.{ ExecutionContext, Future }
-
+import models.JsonFormat.{ analyzerWrites, dataActifactReads, jobWrites }
+import models.{ DataArtifact, FileArtifact }
 import play.api.libs.json.{ JsObject, JsString, Json }
 import play.api.mvc.{ Action, AnyContent, Controller, Request }
-
-import models.{ DataArtifact, FileArtifact }
-import models.JsonFormat.{ analyzerWrites, dataActifactReads, jobWrites }
 import services.{ AnalyzerSrv, JobSrv }
+
+import scala.concurrent.{ ExecutionContext, Future }
 
 class AnalyzerCtrl @Inject() (
     analyzerSrv: AnalyzerSrv,
@@ -38,7 +36,7 @@ class AnalyzerCtrl @Inject() (
   private[controllers] def readFileArtifact(request: Request[AnyContent]) = {
     for {
       parts ← request.body.asMultipartFormData
-      filePart ← parts.file("data").headOption
+      filePart ← parts.file("data")
       attrList ← parts.dataParts.get("_json")
       attrStr ← attrList.headOption
       attr ← Json.parse(attrStr).asOpt[JsObject]
@@ -47,11 +45,12 @@ class AnalyzerCtrl @Inject() (
       ("filename" → JsString(filePart.filename)))
   }
 
-  def analyze(analyzerId: String) = Action.async { request ⇒
+  def analyze(analyzerId: String): Action[AnyContent] = Action.async { request ⇒
     readDataArtifact(request)
       .orElse(readFileArtifact(request))
       .map { artifact ⇒
-        jobSrv.create(artifact, analyzerId)
+        analyzerSrv.analyze(analyzerId, artifact)
+          //jobSrv.create(artifact, analyzerId)
           .map(j ⇒ Ok(Json.toJson(j)))
       }
       .getOrElse(Future.successful(BadRequest("???")))