+++ app/services/JobSrv.scala
@@ -1,79 +1,84 @@
 package services
 
 import java.util.Date
-
 import javax.inject.{ Inject, Named }
-
-import scala.annotation.implicitNotFound
-import scala.concurrent.{ ExecutionContext, Future, Promise }
-import scala.concurrent.duration.{ DurationInt, FiniteDuration }
-import scala.concurrent.duration.Duration
-import scala.concurrent.duration.Duration.Infinite
-import scala.util.Random
 
 import akka.actor.{ Actor, ActorRef, ActorSystem, actorRef2Scala }
 import akka.pattern.ask
 import akka.util.Timeout
-
+import models._
 import play.api.{ Configuration, Logger }
-import play.api.libs.json.{ JsString, JsValue }
 
-import models.{ Analyzer, Artifact, Job, JobStatus }
+import scala.concurrent.duration.Duration.Infinite
+import scala.concurrent.duration.{ Duration, DurationInt, FiniteDuration }
+import scala.concurrent.{ ExecutionContext, Future, Promise }
+import scala.util.Random
 
 class JobSrv @Inject() (
-    analyzerSrv: AnalyzerSrv,
     @Named("JobActor") jobActor: ActorRef,
     implicit val ec: ExecutionContext,
     implicit val system: ActorSystem) {
-  import JobActor._
+
+  import services.JobActor._
+
   implicit val timeout = Timeout(5.seconds)
 
   def list(dataTypeFilter: Option[String], dataFilter: Option[String], analyzerFilter: Option[String], start: Int, limit: Int): Future[(Int, Seq[Job])] = {
     (jobActor ? ListJobs(dataTypeFilter, dataFilter, analyzerFilter, start, limit)).map {
       case JobList(total, jobs) ⇒ total → jobs
-      case _                    ⇒ sys.error("TODO")
+      case m                    ⇒ throw UnexpectedError(s"JobActor.list replies with unexpected message: $m (${m.getClass})")
     }
   }
 
   def get(jobId: String): Future[Job] = (jobActor ? GetJob(jobId)).map {
     case j: Job      ⇒ j
-    case JobNotFound ⇒ sys.error("job not found")
-    case _           ⇒ sys.error("TODO")
+    case JobNotFound ⇒ throw JobNotFoundError(jobId)
+    case m           ⇒ throw UnexpectedError(s"JobActor.GetJob replies with unexpected message: $m (${m.getClass})")
   }
 
-  def create(artifact: Artifact, analyzerId: String): Future[Job] = {
-    analyzerSrv.get(analyzerId)
-      .map { analyzer ⇒ create(artifact, analyzer) }
-      .getOrElse(Future.failed(new Exception("analyzer not found")))
-  }
-
-  def create(artifact: Artifact, analyzer: Analyzer): Future[Job] =
-    (jobActor ? CreateJob(artifact, analyzer)) map {
-      case j: Job ⇒ j
-      case _      ⇒ sys.error("TODO")
+  def create(analyzer: Analyzer, artifact: Artifact, report: Future[Report]): Future[Job] = {
+    (jobActor ? CreateJob(artifact, analyzer, report)) map {
+      case job: Job ⇒ job
+      case m        ⇒ throw UnexpectedError(s"JobActor.CreateJob replies with unexpected message: $m (${m.getClass})")
     }
+  }
+
+  //  @deprecated
+  //  def create(artifact: Artifact, analyzerId: String): Future[Job] = {
+  //    analyzerSrv.get(analyzerId)
+  //      .map { analyzer ⇒ create(artifact, analyzer) }
+  //      .getOrElse(Future.failed(new Exception("analyzer not found")))
+  //  }
+  //
+  //  @deprecated
+  //  def create(artifact: Artifact, analyzer: Analyzer): Future[Job] =
+  //    (jobActor ? CreateJob(artifact, analyzer)) map {
+  //      case j: Job ⇒ j
+  //      case _      ⇒ sys.error("TODO")
+  //    }
 
   def remove(jobId: String): Future[Unit] = {
     (jobActor ? RemoveJob(jobId)).map {
       case JobRemoved  ⇒ ()
-      case JobNotFound ⇒ sys.error("job not found")
-      case _           ⇒ sys.error("TODO")
+      case JobNotFound ⇒ throw JobNotFoundError(jobId)
+      case m           ⇒ throw UnexpectedError(s"JobActor.RemoveJob replies with unexpected message: $m (${m.getClass})")
     }
   }
 
-  def waitReport(jobId: String, atMost: Duration): Future[(JobStatus.Type, JsValue)] = {
-    val statusResult = get(jobId)
-      .flatMap(_.report)
-      .map((JobStatus.Success, _))
-      .recover { case error ⇒ (JobStatus.Failure, JsString(error.getMessage)) }
-
-    atMost match {
-      case _: Infinite ⇒ statusResult
-      case duration: FiniteDuration ⇒
-        val prom = Promise[(JobStatus.Type, JsValue)]()
-        val timeout = system.scheduler.scheduleOnce(duration) { prom.success((JobStatus.Failure, JsString("Timeout"))); () }
-        statusResult onComplete { case _ ⇒ timeout.cancel() }
-        Future.firstCompletedOf(List(statusResult, prom.future))
+  def waitReport(jobId: String, atMost: Duration): Future[Job] = {
+    get(jobId).flatMap { job ⇒
+      val finishedJob = job.report.map(_ ⇒ job)
+      atMost match {
+        case _: Infinite ⇒ finishedJob
+        case duration: FiniteDuration ⇒
+          val prom = Promise[Job]()
+          val timeout = system.scheduler.scheduleOnce(duration) {
+            prom.success(job)
+            ()
+          }
+          finishedJob.onComplete(_ ⇒ timeout.cancel())
+          Future.firstCompletedOf(List(finishedJob, prom.future))
+      }
     }
   }
 }
@@ -83,7 +88,7 @@ object JobActor {
   case class JobList(total: Int, jobs: Seq[Job])
   case class GetJob(jobId: String)
   case object JobNotFound
-  case class CreateJob(artifact: Artifact, analyzer: Analyzer)
+  case class CreateJob(artifact: Artifact, analyzer: Analyzer, report: Future[Report])
   case class RemoveJob(jobId: String)
   case object JobRemoved
   case object JobCleanup
@@ -95,7 +100,8 @@ class JobActor(
     analyzerSrv: AnalyzerSrv,
     implicit val ec: ExecutionContext) extends Actor {
 
-  import JobActor._
+  import services.JobActor._
+
   @Inject def this(
     configuration: Configuration,
     analyzerSrv: AnalyzerSrv,
@@ -127,8 +133,8 @@ class JobActor(
       val filteredJobs = jobs.filter(j ⇒
         dataTypeFilter.fold(true)(j.artifact.dataTypeFilter) &&
           dataFilter.fold(true)(j.artifact.dataFilter) &&
-          analyzerFilter.fold(true)(j.analyzerId.contains))
-      sender ! JobList(filteredJobs.size, filteredJobs.drop(start).take(limit))
+          analyzerFilter.fold(true)(j.analyzer.id.contains))
+      sender ! JobList(filteredJobs.size, filteredJobs.slice(start, start + limit))
     case GetJob(jobId) ⇒ sender ! jobs.find(_.id == jobId).getOrElse(JobNotFound)
     case RemoveJob(jobId) ⇒
       removeJob(jobs, jobId) match {
@@ -137,9 +143,9 @@ class JobActor(
           context.become(jobState(j))
         case None ⇒ sender ! JobNotFound
       }
-    case CreateJob(artifact, analyzer) ⇒
+    case CreateJob(artifact, analyzer, report) ⇒
       val jobId = Random.alphanumeric.take(16).mkString
-      val job = Job(jobId, analyzer.id, artifact, analyzer.analyze(artifact))
+      val job = Job(jobId, analyzer, artifact, report)
       sender ! job
       context.become(jobState(job :: jobs))
     case JobCleanup if jobLifeTime.isInstanceOf[FiniteDuration] ⇒
@@ -148,5 +154,5 @@ class JobActor(
       context.become(jobState(jobs.takeWhile(_.date after limitDate)))
   }
 
-  override def receive = jobState(Nil)
+  override def receive: Receive = jobState(Nil)
 }