+++ ui/app/scripts/components/app-container.component.js
@@ -4,7 +4,17 @@
         .directive('appContainer', function() {
             return {
                 restrict: 'E',
-                templateUrl: 'views/components/app-container.component.html'                
+                templateUrl: 'views/components/app-container.component.html'                ,
+                controller: function(VersionSrv) {
+                    var self = this;
+                    self.config = {};
+
+                    VersionSrv.get()
+                        .then(function(response) {
+                            self.config = response.data;
+                        });
+                },
+                controllerAs: '$vm'
             };
         });
 })();